Notes from Buddy2
The code itself is fine and the application (both tabs) open and run properly. I don't see any code for the layout portion of the assignment and I am unable to get my tablet generator to run so I can't test the proportions. Personal thoughts about the layout and design; the lading page is pretty good... maybe either make the image a little larger or just bring up the email and phone number to fill the white space. I don't think that you really need the image on the second tab. The image overlaps your text. You should also probably reduce the amount of text on the second tab (or probably just remove the image) so you don't have the overlapping problem or making the viewer have to scroll. Another thought, you may think about offsetting your image or text with a contrasting colored box (just to draw the viewers eye).